refactor: Remove AuthDataUseCase
