+++ halu5071/app/src/main/java/io/moatwel/github/presentation/di/UseCaseModule.kt
@@ -25,11 +25,8 @@ package io.moatwel.github.presentation.di
 
 import dagger.Module
 import dagger.Provides
-import io.moatwel.github.domain.repository.AuthDataRepository
 import io.moatwel.github.domain.repository.UserRepository
-import io.moatwel.github.domain.usecase.AuthDataUseCase
 import io.moatwel.github.domain.usecase.UserUseCase
-import javax.inject.Singleton
 
 @Module
 class UseCaseModule {
@@ -37,15 +34,5 @@ class UseCaseModule {
   @Provides
   fun provideUserUseCase(userRepository: UserRepository): UserUseCase {
     return UserUseCase(userRepository)
-  }
-
-  /**
-   *  This AuthDataUseCase is provided as Singleton.
-   *  Because this usecase manage user auth data whole application lifecycle.
-   */
-  @Provides
-  @Singleton
-  fun provideAuthDataUseCase(authDataRepository: AuthDataRepository): AuthDataUseCase {
-    return AuthDataUseCase(authDataRepository)
   }
 }
